Añadiendo products

+++ src/main/java/aiss/api/resources/comparators/ComparatorPriceProduct.java
@@ -8,7 +8,7 @@ public class ComparatorPriceProduct implements Comparator<Product>{
 
 	@Override
 	public int compare(Product o1, Product o2) {
-		return Integer.valueOf(o1.getPrice()).compareTo(Integer.valueOf(o2.getPrice()));
+		return Double.valueOf(o1.getPrice()).compareTo(Double.valueOf(o2.getPrice()));
 	}
 
 }
